uboot: Partition table: Inc userdata. Remove media
Existing media partition is not being used. Removing
this partition and increasing size of userdata partition
where are the media files are stored.

Fixed issue when partition size is greater than 4GB.
The sector size was being read into a u32 instead of u64.
While this will work for smaller sectors, it will not
work for sector sizes larger than 4GB.

+++ cpu/omap4/sparse.c
@@ -49,7 +49,7 @@ int mmc_compare(unsigned mmcc, unsigned char *src, unsigned sector, unsigned len
 }
 
 
-int _unsparse(unsigned char *source, u32 sector, u32 section_size,
+int _unsparse(unsigned char *source, u32 sector, u64 section_size,
 	      unsigned mmcc,
 	      int (*WRITE)(unsigned mwcc, unsigned char *src,
 			   unsigned sector, unsigned len))
@@ -146,7 +146,8 @@ int _unsparse(unsigned char *source, u32 sector, u32 section_size,
 	return 0;
 }
 
-u8 do_unsparse(unsigned char *source, u32 sector, u32 section_size, char *slot_no)
+u8 do_unsparse(unsigned char *source, u32 sector,
+					u64 section_size, char *slot_no)
 {
 	unsigned mmcc = simple_strtoul(slot_no, NULL, 16);
 	if (_unsparse(source, sector, section_size, mmcc, mmc_write))
